Split into Board and View, fix init items bug
