+++ src/Items/List/components/TableView.tsx
@@ -1,16 +1,14 @@
 import AddIcon from "@mui/icons-material/Add";
 // import { Stack } from "@mui/material";
-import Box from "@mui/material/Box";
 import Fab from "@mui/material/Fab";
 import Paper from "@mui/material/Paper";
 import { useTheme } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
-import React, { useState } from "react";
+import React from "react";
 import { useNavigate } from "react-router";
-import type { DropResult } from "react-beautiful-dnd";
-import DraggableList from "../../../DragDrop/DraggableList";
-import { reorder } from "../../../helpers/helpers";
+
 import { useItems } from "../../state/ItemContext";
+import { Board } from "./Board";
 // import type { Item } from "../../types";
 // import { ItemTableRow } from "./ItemTableRow";
 
@@ -21,22 +19,14 @@ const fabStyle = {
 };
 
 export function TableView() {
-  const { items: fetchedItems } = useItems();
-  const [items, setItems] = useState(fetchedItems);
+  const { items } = useItems();
+  console.log(items);
+
   const theme = useTheme();
   const navigate = useNavigate();
 
   const onAddItem = () => {
     navigate("/create");
-  };
-
-  const onDragEnd = ({ destination, source }: DropResult) => {
-    // dropped outside the list
-    if (!destination) return;
-
-    const newItems = reorder(items, source.index, destination.index);
-
-    setItems(newItems);
   };
 
   return (
@@ -45,32 +35,8 @@ export function TableView() {
         My items
       </Typography>
       <Paper>
-        <DraggableList items={items} onDragEnd={onDragEnd} />
+        <Board fetchedItems={items} />
       </Paper>
-      <Box sx={{ m: 3 }}>
-        {/* <Stack spacing={2}> */}
-        {/*   {items.map((item: Item) => ( */}
-        {/*     <ItemTableRow item={item} key={item.id} /> */}
-        {/*   ))} */}
-        {/* </Stack> */}
-
-        {/* <TableContainer component={Paper}> */}
-        {/*  <Table aria-label="simple table"> */}
-        {/*    <TableHead> */}
-        {/*      <TableRow> */}
-        {/*        <TableCell>Title</TableCell> */}
-        {/*        <TableCell>Author</TableCell> */}
-        {/*        <TableCell>Description</TableCell> */}
-        {/*        <TableCell align="right">Progress</TableCell> */}
-        {/*        <TableCell> </TableCell> */}
-        {/*      </TableRow> */}
-        {/*    </TableHead> */}
-        {/*    <TableBody> */}
-        {/*      */}
-        {/*    </TableBody> */}
-        {/*  </Table> */}
-        {/* </TableContainer> */}
-      </Box>
 
       <Fab color="success" aria-label="add" sx={fabStyle}>
         <AddIcon onClick={onAddItem} />
